cleanup, small changes changed to new arctrl 2.0.0; exported file and folder creation and deletion to separate area; removed retries for 400 status codes
