+++ requirements.txt
@@ -13,5 +13,5 @@ httpx~=0.24.1
 itsdangerous==2.1.2
 python-multipart~=0.0.6
 fsspreadsheet~=6.1.1
-arctrl~=1.2.0
+arctrl~=2.0.0
 pdf2image~=1.17.0
